Installs GO v1.16.5

+++ golang_installer.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
-sudo apt install tar -y
-cd $HOME
-wget https://golang.org/dl/go1.16.5.linux-amd64.tar.gz
-rm -rf /usr/local/go
-tar -C /usr/local -xzf go1.16.5.linux-amd64.tar.gz
-rm go1.16.5.linux-amd64.tar.gz
-echo "export PATH=$PATH:/usr/local/go/bin:$HOME/go/bin" >> ~/.bash_profile
-source ~/.bash_profile
+version="1.16.5"
+if ! go version | grep -q $version; then
+	sudo apt install tar -y
+	cd $HOME
+	wget "https://golang.org/dl/go$version.linux-amd64.tar.gz"
+	rm -rf /usr/local/go
+	tar -C /usr/local -xzf "go$version.linux-amd64.tar.gz"
+	rm go1.16.5.linux-amd64.tar.gz
+	echo "export PATH=$PATH:/usr/local/go/bin:$HOME/go/bin" >> ~/.bash_profile
+	source ~/.bash_profile
+fi
